moves views to views folder
