+++ DailyStatus.UI/View/StatusWindow.xaml
@@ -1,4 +1,4 @@
-﻿<Window x:Class="DailyStatus.UI.MainWindow"
+﻿<Window x:Class="DailyStatus.UI.View.StatusWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
@@ -6,7 +6,7 @@
         xmlns:lvc="clr-namespace:LiveCharts.Wpf;assembly=LiveCharts.Wpf"
         xmlns:local="clr-namespace:DailyStatus.UI"
         mc:Ignorable="d"
-        Title="MainWindow" Height="110" Width="230">
+        Title="Status" Height="150" Width="250">
     <Grid>
         <lvc:AngularGauge FromValue="-16" ToValue="16" 
                           LabelsStep="4" TicksStep="4" Wedge="180"
