+++ DailyStatus.UI/View/StatusWindow.xaml.cs
@@ -1,7 +1,4 @@
-﻿using DailyStatus.UI.ViewModel;
-using LiveCharts;
-using LiveCharts.Wpf;
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Text;
@@ -13,20 +10,18 @@ using System.Windows.Documents;
 using System.Windows.Input;
 using System.Windows.Media;
 using System.Windows.Media.Imaging;
-using System.Windows.Navigation;
 using System.Windows.Shapes;
 
-namespace DailyStatus.UI
+namespace DailyStatus.UI.View
 {
     /// <summary>
-    /// Interaction logic for MainWindow.xaml
+    /// Interaction logic for StatusWindow.xaml
     /// </summary>
-    public partial class MainWindow : Window
+    public partial class StatusWindow : Window
     {
-        public MainWindow()
+        public StatusWindow()
         {
             InitializeComponent();
-            DataContext = new MainWindowViewModel();
         }
     }
 }
